Relocalized all methods to StandardMethods class
