+++ app/src/main/java/com/example/wise_batteryexam2023/methods/Test.kt
@@ -1,0 +1,5 @@
+package com.example.wise_batteryexam2023.methods
+
+import android.content.Context
+
+class Test(private val applicationContext: Context)
